requirements.txt
Django==2.0.2
djangorestframework==3.7.7
Pillow==5.0.0
